+++ packages/cli/src/main.ts
@@ -29,7 +29,7 @@ export async function main(argv = process.argv): Promise<void> {
     .action(runChecksCommand);
 
   cli.help();
-  cli.version("0.2.5");
+  cli.version("0.3.0");
   cli.parse(argv, { run: false });
   await cli.runMatchedCommand();
 }